moved axios.post call, handleSubmit, handleChange to App.js to handle context API, all is working and adding to list, need to look into why you have to refresh to update list and finish styling
