+++ smurfs/src/components/App.js
@@ -11,11 +11,21 @@ import Smurf from '../Smurfs/Smurf';
 import SmurfForm from "../Smurfs/SmurfForm";
 
 import "./App.css";
+import { AddSmurfContext } from "../context/AddSmurfContext";
 
 const App = () => {
+  // initial state
   const [smurfs, setSmurfs] = useState([]);
-  console.log('APP', smurfs)
+  // console.log('APP', smurfs)
+  // adding new smurfs to state
+  const [smurf, setSmurf] = useState({
+    name: '',
+    age: '',
+    height: '',
+    id: ''
+  })
 
+  // axios call for getting smurfs to render
   useEffect(() => {
     axios
     .get('http://localhost:3333/smurfs')
@@ -28,19 +38,42 @@ const App = () => {
     })
   }, [])
 
-  // const addSmurfs = () => {
-  //   setSmurfs()
-  // }
+  const handleChange = e => {
+    setSmurf({
+      ...smurf,
+      [e.target.name]: e.target.value
+    })
+  }
+
+  const handleSubmit = e => {
+    e.preventDefault()
+
+    axios
+      .post('http://localhost:3333/smurfs', smurf)
+      .then(result => {
+        setSmurf({
+          name: '',
+          age: '',
+          height: ''
+        })
+        console.log('Smurf has been added', result.data)  
+      })
+      .catch(error => {
+        console.log('POST ERROR', error)
+      })
+  }
 
   return (
     <SmurfContext.Provider value={smurfs}>
-      <div className="App">
-        <Header />
+      <AddSmurfContext.Provider value={{smurf, handleChange, handleSubmit}}>
+        <div className="App">
+          <Header />
 
-        {/* Routes */}
-        <Route exact path='/smurfs' component={Smurf} />
-        <Route exact path='/addsmurf' component={SmurfForm} />
-      </div>
+          {/* Routes */}
+          <Route exact path='/smurfs' component={Smurf} />
+          <Route exact path='/addsmurf' component={SmurfForm} />
+        </div>
+      </AddSmurfContext.Provider>
     </SmurfContext.Provider>
   );
 }
