+++ smurfs/src/context/AddSmurfContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const AddSmurfContext = createContext();
